server/requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
olefile==0.45.1
Pillow==4.3.0
PyJWT==1.6.1
pytz==2018.3
